requirements.txt
opencv-python 
numpy 
pandas 
supervision 
torch 
torchvision 
ultralytics 
plotly 
colorama 
python-telegram-bot 
python-dotenv
agno
google-genai
groq
sqlalchemy
fastapi
uvicorn
supabase
